backend/app/llm.py
import google.generativeai as genai
import os
from dotenv import load_dotenv

load_dotenv()

genai.configure(api_key=os.getenv("GEMINI_API_KEY"))

# ✅ Updated model
MODEL = genai.GenerativeModel("gemini-2.5-flash")

SYSTEM_PROMPT = """
You are an AI customer support agent.

You must:
- Understand user intent
- Extract order_id if present
- Decide which tool to call

Available tools:
- get_order(order_id)
- get_customer(email)
- check_refund_eligibility(order_id)
- escalate_ticket(reason)

IMPORTANT:
Respond ONLY in JSON format:

{
  "action": "tool_name OR final",
  "parameters": {},
  "response": ""
}

Rules:
- NEVER guess data
- ALWAYS call tools when needed
- DO NOT return raw tool data
"""