backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
google-generativeai
python-dotenv
pydantic